add BaseFilterManager
